RDKBDEV-2732 : Fix error after enabling -Werror and -Wall in RdkGponManager.
Reason for change:
Fix error after enabling -Werror and -Wall in RdkGponManager.

Test Procedure: Performed GponManager Sanity test.
Risks: None.

+++ source/GponManager/gponmgr_controller.c
@@ -28,7 +28,7 @@ ANSC_STATUS GponMgr_subscribe_hal_events(void)
 {
     ANSC_STATUS ret = ANSC_STATUS_FAILURE;
     char strName[JSON_MAX_STR_ARR_SIZE]={0};
-    int idx = 0;
+    ULONG idx = 0;
 
     GPON_DML_DATA* pGponDmlData = GponMgrDml_GetData_locked();
     if(pGponDmlData != NULL)
